fix workflows and docker file

+++ up.sh
@@ -1,17 +1,14 @@
 #!/bin/bash
 
-# Build java app
-#cd redis-java-app
-#  #mvn -B dependency:go-offline
-#  mvn clean install -DskipTests=true
-#STATUS=$?
-#if [ $STATUS != 0 ]; then
-#  echo "Maven build failed."
-#  exit 1
-#fi
-#cd ..
+#Build java app
+./backend/gradlew build bootJar -p backend
+STATUS=$?
+if [ $STATUS != 0 ]; then
+  echo "Maven build failed."
+  exit 1
+fi
 
 docker-compose down -v
 # Deploy services using docker-compose
-docker-compose up --build --remove-orphans
+docker-compose up --build --remove-orphans -d
 exit 0
